requirements.txt
celery==4.1.0
Flask==0.12.2
boto3==1.4.7
PyYAML==3.12
pycurl==7.43.0
